requirements.txt
Flask==3.0.0
tensorflow==2.15.0
numpy==1.26.2
pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.6.1
gunicorn==21.2.0